+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Home: Renamed "Update ESI data" button to "Update corporation".
 - Added "player" option to "update-player-groups" command.
 - The "update player groups" function now also synchronizes the manager role, just in case there was an error before.
+- Improved performance of permission check for "/api/user/player/{id}/characters".
 
 ## 1.17.1